requirements.txt
echo "streamlit
langchain
langchain-community
langchain-openai
langchain-huggingface
langchain-classic
faiss-cpu
pypdf
python-dotenv
sentence-transformers" > requirements.txt